Ehhh Not Much
- added a new enemy (bumbler)
- added new powerup (damage boost)
- rearranged files
- powerup pickup notification
- rooms 50-51 (courtesy of sam uwe alws)
- room 52 (courtesy of eric zheng)
- room 53 (courtesy of bryan yao)

+++ games4jeffpackage/EnemyShot.java
@@ -8,11 +8,11 @@ import java.awt.Rectangle;
 public class EnemyShot extends GameThing{
 
 	private float angle;
-	private int damage;
+	private float damage;
 	private Handler handler;
 	private int timer = 0;
 
-	public EnemyShot(float x, float y, String id, float angle, int damage, int range, Handler handler) {
+	public EnemyShot(float x, float y, String id, float angle, float damage, int range, Handler handler) {
 		super(x, y, id);
 		this.angle = angle;
 		this.damage = damage;
@@ -48,7 +48,7 @@ public class EnemyShot extends GameThing{
 		return new Rectangle((int)x, (int)y, (int)width, (int)height);
 	}
 
-	public int getDamage(){
+	public float getDamage(){
 		return damage;
 	}
